// libs/neural-ng/button/button.types.ts
import type { NeuralBadgePosition } from '@neural-ng/core/badge';

export type NeuralButtonBadgePosition = NeuralBadgePosition;

export type NeuralButtonIconPosition = 'start' | 'end';

export type NeuralButtonSize = 'small' | 'medium' | 'large';
export type NeuralButtonIconSize = 'small' | 'medium' | 'large';

export type NeuralButtonGroupOrientation = 'horizontal' | 'vertical';

export type NeuralButtonSeverity =
  | 'primary'
  | 'secondary'
  | 'neutral'
  | 'info'
  | 'success'
  | 'warning'
  | 'error';